Fix xrServer event handler signature mismatches

+++ src/xrGame/xrServer_process_event_reject.cpp
@@ -3,7 +3,7 @@
 #include "xrserver_objects.h"
 
 bool xrServer::Process_event_reject(NET_Packet& P, const ClientID sender, const u32 time, const u32 id_parent,
-                                    const u16 id_entity, bool send_message)
+                                    const u32 id_entity, bool send_message)
 {
 	// Parse message
 	CSE_Abstract* e_parent = game->get_entity_from_eid(id_parent);
